Move lens related code from vapors.data to vapors.lens
