+++ core/src/main/scala/com/rallyhealth/vapors/core/data/Fact.scala
@@ -2,6 +2,7 @@ package com.rallyhealth.vapors.core.data
 
 import cats.Order
 import cats.syntax.all._
+import com.rallyhealth.vapors.core.lens.{DataPath, NamedLens}
 
 sealed abstract class Fact {
   type Value
