+++ core/src/main/scala/com/rallyhealth/vapors/core/data/FactType.scala
@@ -1,6 +1,7 @@
 package com.rallyhealth.vapors.core.data
 
 import cats.Order
+import com.rallyhealth.vapors.core.lens.{NamedLens, ValidDataPathKey}
 
 import scala.reflect.ClassTag
 import scala.reflect.runtime.universe.TypeTag
