+++ core/src/main/scala/com/rallyhealth/vapors/core/lens/DataPath.scala
@@ -1,4 +1,4 @@
-package com.rallyhealth.vapors.core.data
+package com.rallyhealth.vapors.core.lens
 
 import cats.Show
 import cats.data.NonEmptySet
